+++ src/js/node/_http_server.ts
@@ -233,19 +233,6 @@ function emitListenErrorNextTick(self, err) {
   self.emit("error", err);
 }
 
-// Node.js only requests a client certificate when `requestCert: true`.
-// The uSockets SSL context treats `ca` alone as "verify peer", so without
-// these two flags an `https.Server({ ca })` would reject every client that
-// doesn't present a cert. Mirror tls.Server (net.ts): default `requestCert`
-// to false and, when not requesting, force `rejectUnauthorized` to false so
-// the CA is loaded into the trust store without requiring a client cert.
-function normalizeServerTls(tls) {
-  const requestCert = !!tls.requestCert;
-  tls.requestCert = requestCert;
-  tls.rejectUnauthorized = requestCert ? tls.rejectUnauthorized !== false : false;
-  return tls;
-}
-
 // Node registers connectionListener on every http.Server so `server.emit("connection", socket)`
 // works for foreign Duplex sockets. The native listener handles its own sockets end to end;
 // this picks up the rest. https://github.com/nodejs/node/blob/main/lib/_http_server.js
@@ -259,95 +246,9 @@ function connectionListener(this: Server, socket) {
   });
 }
 
-// Always returns a config, so https.Server speaks TLS without key/cert like Node's tls.Server.
-function serverTlsOptions(options) {
-  const tlsHelpers = require("internal/tls");
-
-  // Node's https.Server accepts PKCS#12 bundles (pfx [+ passphrase]); fold
-  // them into plain key/cert/ca so the native TLS config sees PEM material.
-  let tlsOptions = options;
-  if (options.pfx) {
-    tlsOptions = tlsHelpers.processPfxOptions(options);
-  }
-
-  let cert = tlsOptions.cert;
-  if (cert) {
-    tlsHelpers.throwOnInvalidTLSArray("options.cert", cert);
-  }
-
-  let key = tlsOptions.key;
-  if (key) {
-    tlsHelpers.throwOnInvalidTLSArray("options.key", key);
-  }
-
-  let ca = tlsOptions.ca;
-  // PKCS#12-embedded CAs extend the trust set; the server path hands raw
-  // {key, cert, ca} to the native config and has no addCACert hook, so fold
-  // them into `ca` (mirrors tls.Server.setSecureContext).
-  const pfxExtraCAs = tlsOptions._pfxExtraCACerts;
-  if (pfxExtraCAs?.length) {
-    ca = ca == null ? pfxExtraCAs : $isArray(ca) ? [...ca, ...pfxExtraCAs] : [ca, ...pfxExtraCAs];
-  }
-  if (ca) {
-    tlsHelpers.throwOnInvalidTLSArray("options.ca", ca);
-  }
-
-  let passphrase = options.passphrase;
-  if (passphrase && typeof passphrase !== "string") {
-    throw $ERR_INVALID_ARG_TYPE("options.passphrase", "string", passphrase);
-  }
-
-  let serverName = options.servername;
-  if (serverName && typeof serverName !== "string") {
-    throw $ERR_INVALID_ARG_TYPE("options.servername", "string", serverName);
-  }
-
-  let secureOptions = options.secureOptions || 0;
-  if (secureOptions && typeof secureOptions !== "number") {
-    throw $ERR_INVALID_ARG_TYPE("options.secureOptions", "number", secureOptions);
-  }
-
-  const { validateSecureProtocol, secureProtocolToVersionRange, tlsStringToProtocolVersion } = tlsHelpers;
-  // Translate minVersion/maxVersion/secureProtocol into the integer
-  // protocol range the native layer applies (secureProtocol wins, like
-  // Node's SecureContext::Init); 0 keeps the native defaults.
-  validateSecureProtocol(options.secureProtocol);
-  let minVersion, maxVersion;
-  const range = secureProtocolToVersionRange(options.secureProtocol);
-  if (range) {
-    minVersion = range[0];
-    maxVersion = range[1];
-  } else {
-    minVersion = tlsStringToProtocolVersion(options.minVersion);
-    maxVersion = tlsStringToProtocolVersion(options.maxVersion);
-  }
-  return normalizeServerTls({
-    serverName,
-    key,
-    cert,
-    ca,
-    passphrase,
-    secureOptions,
-    minVersion,
-    maxVersion,
-    ciphers: typeof options.ciphers === "string" && options.ciphers ? options.ciphers : undefined,
-    requestCert: options.requestCert,
-    rejectUnauthorized: options.rejectUnauthorized,
-  });
-}
-
-// Like Node, http.Server ignores key/cert/ca/pfx. Only HttpsServer reads them.
+// Like Node, http.Server ignores key/cert/ca/pfx. https.Server (https.ts) sets its own TLS config.
 function Server(options, callback): void {
   if (!(this instanceof Server)) return new Server(options, callback);
-  return initServer.$call(this, options, callback, false);
-}
-
-function HttpsServer(options, callback): void {
-  if (!(this instanceof HttpsServer)) return new HttpsServer(options, callback);
-  return initServer.$call(this, options, callback, true);
-}
-
-function initServer(options, callback, isTls: boolean) {
   EventEmitter.$call(this);
   this.on("listening", setupConnectionsTracking);
   this.on("connection", connectionListener);
@@ -375,9 +276,6 @@ function initServer(options, callback, isTls: boolean) {
     validateObject(options, "options");
     options = { ...options };
   }
-  if (isTls) {
-    this[tlsSymbol] = serverTlsOptions(options);
-  }
 
   this[optionsSymbol] = options;
   storeHTTPOptions.$call(this, options);
@@ -386,7 +284,6 @@ function initServer(options, callback, isTls: boolean) {
   return this;
 }
 $toClass(Server, "Server", EventEmitter);
-$toClass(HttpsServer, "Server", Server);
 
 Server.prototype[kIncomingMessage] = undefined;
 
@@ -560,7 +457,7 @@ Server.prototype.listen = function () {
 
       const otherTLS = arg0.tls;
       if (otherTLS && $isObject(otherTLS)) {
-        tls = normalizeServerTls({ ...otherTLS });
+        tls = require("internal/tls").normalizeServerTls({ ...otherTLS });
       }
     } else if (typeof arg0 === "string" && !(Number(arg0) >= 0)) {
       // (path[...][, cb])
@@ -3822,7 +3719,6 @@ function storeHTTPOptions(options) {
 
 export default {
   Server,
-  HttpsServer,
   ServerResponse,
   kConnectionsCheckingInterval,
 };
